Fix corruption when FD_SETSIZE larger than expected
The expected FD_SETSIZE for Select is currently hardcoded to be 1024.  While common, this is not required, and many servers will have this limit increased.  With the current hardcoded value, if a file descriptor comes in beyond the limit, we end up corrupting memory.

This commit fixes it by making the limit in managed code dynamic, supplied via a call to the native shim to return the actual value of FD_SETSIZE so that the managed and native expectations match.

(cherry picked from commit a0efed6cac8eb2528938f5e7b15fa1205cc98095)

Conflicts:
	src/Native/System.Native/pal_networking.cpp
	src/Native/System.Native/pal_networking.h